smart-land-guard-backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic[email]==2.7.1
pydantic-settings==2.3.1
cloudinary==1.40.0
twilio==9.0.5
sendgrid==6.11.0
weasyprint==62.1
jinja2==3.1.4
numpy==1.26.4
Pillow==10.3.0
aiofiles==23.2.1
passlib[bcrypt]==1.7.4
